requirements.txt
amqp==1.4.9
anyjson==0.3.3
appnope==0.1.0
asn1crypto==0.24.0
attrs==18.1.0
Automat==0.7.0
backcall==0.1.0
bcrypt==3.1.4
billiard==3.3.0.23
celery==3.1.26.post2
certifi==2018.4.16
cffi==1.11.5
chardet==3.0.4
click==6.7
constantly==15.1.0
cryptography==2.2.2
cssselect==1.0.3
decorator==4.3.0
defusedxml==0.5.0
Django==2.1
django-ckeditor==5.5.0
django-debug-toolbar==1.9.1
django-js-asset==1.1.0
django-localflavor==2.1
django-taggit==0.22.2
django-tinymce==2.7.0
django-widget-tweaks==1.4.2
Faker==0.9.0
Flask==1.0.2
Flask-Login==0.4.1
future==0.16.0
hyperlink==18.0.0
idna==2.7
incremental==17.5.0
ipython==6.4.0
ipython-genutils==0.2.0
itsdangerous==0.24
jedi==0.12.0
Jinja2==2.10
jsmin==2.2.2
kombu==3.0.37
lxml==4.2.2
MarkupSafe==1.0
parsel==1.4.0
parso==0.2.1
pexpect==4.6.0
pickleshare==0.7.4
Pillow==5.2.0
ply==3.11
prompt-toolkit==1.0.15
ptyprocess==0.5.2
pyasn1==0.4.3
pyasn1-modules==0.2.1
pycparser==2.18
pycurl==7.43.0.2
PyDispatcher==2.0.5
Pygments==2.2.0
pyOpenSSL==18.0.0
pyquery==1.4.0
python-dateutil==2.7.3
pytz==2018.5
PyYAML==3.12
queuelib==1.5.0
requests==2.19.1
Scrapy==1.5.0
service-identity==17.0.0
simplegeneric==0.8.1
six==1.11.0
sqlparse==0.2.4
tblib==1.3.2
text-unidecode==1.2
tornado==4.5.3
traitlets==4.3.2
Twisted==18.4.0
u-msgpack-python==2.5.0
urllib3==1.23
w3lib==1.19.0
wcwidth==0.1.7
Werkzeug==0.14.1
WsgiDAV==2.4.1
zope.interface==4.5.0
